Remove pagination service
